+++ src/resources/routes/admin.php
@@ -22,6 +22,7 @@ Route::resource('property', 'PropertyController');
 Route::resource('order', 'OrderController');
 Route::resource('media', 'MediaController')->only(['update', 'destroy', 'store']);
 Route::resource('carrier', 'CarrierController');
+Route::resource('payment', 'PaymentController')->only(['index']);
 Route::resource('payment-method', 'PaymentMethodController')->parameters(['payment-method' => 'paymentMethod']);
 Route::resource('shipping-method', 'ShippingMethodController')->parameters(['shipping-method' => 'shippingMethod']);
 Route::resource('tax-category', 'TaxCategoryController')->parameters(['tax-category' => 'taxCategory']);